requirements.txt
websocket
pypng
curve25519-donna
pyqrcode
dissononce
